Made the bot not report itself for deleting a suggestion.

+++ Modules/Suggest/got_deleted.py
@@ -15,12 +15,13 @@ class got_deleted(commands.Cog):
 
     @commands.Cog.listener()
     async def on_raw_message_delete(self, payload):
-        guild = self.client.get_guild(payload.guild_id)
-        channel = guild.get_channel(payload.channel_id)
+        if not sql.is_deleted(payload.message_id):
+            guild = self.client.get_guild(payload.guild_id)
+            channel = guild.get_channel(payload.channel_id)
 
-        if payload.channel_id == self.config.get("suggestion_channel"):
-            sql.delete_suggestion(payload.message_id)
-            console.log(f"Suggestion deleted by force: {payload.message_id}.")
+            if payload.channel_id == self.config.get("suggestion_channel"):
+                sql.delete_suggestion(payload.message_id)
+                console.log(f"Suggestion deleted by force: {payload.message_id}.")
 
 def setup(client):
     client.add_cog(got_deleted(client))
